Update Module 04 - Group Lab Defined
This content walks through a group lab for scanning a Windows node, addressing the issue, and then rescanning. This assumes a few things at this time. It is assumed that the Chef Compliance Server is running Inspec 0.10.1 and the node itself is running InSpec 0.10.1. As of writing these slides this is not the case.